test(dolt): wire TestRunDoltSQLHandlesLargeScript into the default lane (refs be-wq49)
Reviewer follow-up on be-go6d (runDoltSQL argv E2BIG fix): the regression
test proving it lived in git_remote_test.go under //go:build integration,
gated by skipIfNoDolt — which checks for the shared containerized Dolt SQL
server (testServerPort). That test never touches the server, only the
local dolt CLI, so the gate was wrong and the test only ever ran on the
non-blocking nightly cron, never in the PR-gating/merge-gating lane.

Extract TestRunDoltSQLHandlesLargeScript + its runDoltSQL helper into a
new file, dolt_sql_large_script_test.go, with no tag and gated on
testutil.RequireDoltBinary(t) instead. Remove the originals from
git_remote_test.go (kept as a same-package call site — setupGitRemote
still calls runDoltSQL) to avoid duplicate symbols when both files
compile together under -tags=integration.

No production code changes: runDoltSQL's fix was already proven correct
in be-go6d's own tdd_red/tdd_green (cc407e186/fceb1c69a). This closes the
CI-wiring gap only.

Verified: test absent from the default lane before (mechanically-verified
red state via a targeted -run + -v check), present and passing after;
default and integration-tagged compiles both stay duplicate-symbol free;
full ./internal/storage/dolt/... default-lane suite still green with no
regression.
