+++ internal/storage/dolt/git_remote_test.go
@@ -141,39 +141,10 @@ func runCmd(t *testing.T, dir string, name string, args ...string) {
 	}
 }
 
-// runDoltSQL executes SQL via `dolt sql` CLI in the given directory. The
-// script is piped over stdin rather than passed as a `-q` argv element:
-// schema.AllMigrationsSQL() is ~134KB, past Linux's per-argv MAX_ARG_STRLEN
-// (131072 bytes), which fails execve with E2BIG when passed as an argument.
-func runDoltSQL(t *testing.T, dir, query string) {
-	t.Helper()
-	cmd := exec.Command("dolt", "sql")
-	cmd.Dir = dir
-	cmd.Stdin = strings.NewReader(query)
-	if output, err := cmd.CombinedOutput(); err != nil {
-		t.Fatalf("dolt sql failed in %s: %v\nQuery: %.200s...\nOutput: %s", dir, err, query, output)
-	}
-}
-
-// TestRunDoltSQLHandlesLargeScript proves runDoltSQL can execute a SQL
-// script larger than Linux's per-argv MAX_ARG_STRLEN (128KiB = 131072
-// bytes). Passing the script as a single argv element (`dolt sql -q
-// <script>`) blows past that limit and fails execve with E2BIG; the real
-// schema init script (schema.AllMigrationsSQL()) is ~134KB and triggers it.
-func TestRunDoltSQLHandlesLargeScript(t *testing.T) {
-	skipIfNoDolt(t)
-
-	dir := t.TempDir()
-	runCmd(t, dir, "dolt", "init")
-
-	script := schema.AllMigrationsSQL()
-	const maxArgStrlen = 131072 // Linux MAX_ARG_STRLEN, in bytes
-	if len(script) <= maxArgStrlen {
-		t.Fatalf("schema script is %d bytes, must exceed MAX_ARG_STRLEN (%d) to exercise the argv-limit path", len(script), maxArgStrlen)
-	}
-
-	runDoltSQL(t, dir, script)
-}
+// runDoltSQL lives in dolt_sql_large_script_test.go (untagged, so it's
+// always in scope here too) rather than in this integration-tagged file,
+// since TestRunDoltSQLHandlesLargeScript there needs it without the
+// integration tag.
 
 // skipIfNoGit skips if git is not available.
 func skipIfNoGit(t *testing.T) {
